#11 - token searcher: move parts of the logic into one module which is responsible for searching
