+++ lib/util/findNextPunctuators.js
@@ -10,31 +10,102 @@
 
 'use strict';
 
-var collectVariableDeclarations = require('./collectVariableDeclarations');
-var getNextObjectAssignment = require('./getNextObjectAssignment');
 var isVariableAssignment = require('./isVariableAssignment');
-
-function findNextExpressionToken(source, node) {
-  return source.getTokenAfter(!isVariableAssignment(node) ? node.parent : node, !isVariableAssignment(node) ? 2 : 1);
-}
-
-function isNextObjectExpression(source, node) {
-  try {
-    var next = findNextExpressionToken(source, node);
-  } catch (e) {
-    return false;
-  }
-
-  if (typeof next === 'undefined') {
-    if (isVariableAssignment(node) && node.declarations.length > 1) {
-      return false;
-    }
-    throw 'Failed to receive next expression!';
-  }
-  return -1 !== ['.', '['].indexOf(next.value);
-}
+var isTernaryOperator = require('./isTernaryOperator');
 
 module.exports = function(node, source, lines) {
+  function isPunctuatorAssignment(token) {
+    return token.type === 'Punctuator' && token.value === '=';
+  }
+
+  function getPunctuators(node, source) {
+    function isMultipleLineDeclaration(node, token) {
+      return node.loc.start.line < token.loc.start.line;
+    }
+
+    function gatherPunctuatorsFromCurrentDeclaration(node, source) {
+      return node.declarations.reduce(function (tokens, subToken, index) {
+        // skip first one and empty declarations
+        if (subToken.init && index > 0) {
+          var punctuator = source.getTokensBetween(subToken.id, subToken.init).shift();
+          if (isPunctuatorAssignment(punctuator) && isMultipleLineDeclaration(node, punctuator)) {
+            tokens.push(punctuator);
+          }
+        }
+        return tokens;
+      }, []);
+    }
+
+    var punctuators = [];
+    if (isVariableAssignment(node) && node.declarations.length > 1) {
+      punctuators = gatherPunctuatorsFromCurrentDeclaration(node, source);
+    }
+
+    return punctuators;
+  }
+
+  function collectVariableDeclarations(node, source) {
+    function isNewVariable(expression) {
+      return expression.type === 'Keyword' && -1 !== ['var', 'let', 'const'].indexOf(expression.value);
+    }
+
+    var isAssignment = isVariableAssignment(node),
+      after = source.getTokenAfter(node.declarations ? node : node.parent, isAssignment ? 0 : 1),
+      punctuators = getPunctuators(node, source);
+
+    // check the next line for further expressions and skip completely in case of empty lines
+    if (typeof after !== 'undefined') {
+      // checks whether a new variable is introduced
+      // or an existing one will be changed.
+      var punctuator = isNewVariable(after) ? source.getTokenAfter(after, 1) : source.getTokenAfter(after);
+
+      if (isPunctuatorAssignment(punctuator)) {
+        punctuators.push(punctuator);
+      }
+    }
+
+    return punctuators;
+  }
+
+  function getNextObjectAssignment(source, node) {
+    function getTokenSearchAmount(node, isObjectAssignment) {
+      return (isObjectAssignment && node.computed)
+        ? 5
+        : (!isObjectAssignment && (node.declarations.length > 1) || isTernaryOperator(node)) ? 3 : 4;
+    }
+
+    try {
+      var isObjectAssignment = !isVariableAssignment(node);
+      return getPunctuators(node, source).concat([
+        source.getTokenAfter(isObjectAssignment ? node.parent : node, getTokenSearchAmount(node, isObjectAssignment))
+      ]);
+    } catch (e) {
+      // if an object assignment is in the last line (like a module.exports expression)
+      // the token searcher may fail
+      return [];
+    }
+  }
+
+  function isNextObjectExpression(source, node) {
+    function findNextExpressionToken(source, node) {
+      return source.getTokenAfter(!isVariableAssignment(node) ? node.parent : node, !isVariableAssignment(node) ? 2 : 1);
+    }
+
+    try {
+      var next = findNextExpressionToken(source, node);
+    } catch (e) {
+      return false;
+    }
+
+    if (typeof next === 'undefined') {
+      if (isVariableAssignment(node) && node.declarations.length > 1) {
+        return false;
+      }
+      throw 'Failed to receive next expression!';
+    }
+    return -1 !== ['.', '['].indexOf(next.value);
+  }
+
   var nextLine = lines[node.loc.start.line];
   if (typeof nextLine !== 'undefined' && nextLine.trim().length !== 0) {
     try {
